Fix bug preventing sample number to be more that 128. New limit is 32768

+++ Examples/FFT_01/FFT_01.ino
@@ -84,7 +84,7 @@ void loop()
   // delay(2000); /* Repeat after delay */
 }
 
-void PrintVector(double *vData, uint8_t bufferSize, uint8_t scaleType)
+void PrintVector(double *vData, uint16_t bufferSize, uint8_t scaleType)
 {
   for (uint16_t i = 0; i < bufferSize; i++)
   {
