+++ Examples/FFT_02/FFT_02.ino
@@ -95,7 +95,7 @@ void loop()
   while(1); /* Run Once */
 }
 
-void PrintVector(double *vData, uint8_t bufferSize, uint8_t scaleType)
+void PrintVector(double *vData, uint16_t bufferSize, uint8_t scaleType)
 {
   for (uint16_t i = 0; i < bufferSize; i++)
   {
